backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
mysql-connector-python==9.1.0
python-dotenv==1.0.1
numpy==2.2.1
opencv-python==4.10.0.84
insightface==0.7.3
onnxruntime==1.20.1
Pillow==11.0.0
email-validator==2.2.0
bcrypt==4.2.1
PyJWT==2.10.1
